+++ syntax/attribute_value.mli
@@ -220,6 +220,10 @@ val script_type : parser
     These parsers are named after the attribute for which they are used. *)
 
 val sandbox : parser
+val blocking : parser
+val ol_type : parser
+val popover : parser
+val command : parser
 val in_ : parser
 val in2 : parser
 val xmlns : parser